feat(degreeworks-scraper): discover and scrape only majors that exist (#213)
This closes #156.

# What was done

We use the data backing the table at https://www.reg.uci.edu/mdsd/ to
gather the list of valid "majors" at UCI. This term is used informally
because it is a somewhat overloaded term. There are actually several
parts to a major:

- The "school", with a unique code, e.g. Social Sciences.
- The "major", with a unique code, e.g. Cognitive Sciences.
- The "degree", with a unique code, e.g. B.S.

The combination of these three (and the correct department, but we don't
need that for reasons explained in the code) forms the colloquial notion
of a "major". Note that this also applies to graduate programs.

The identifiers describing schools and majors in this table align
exactly with DegreeWorks. However, the degree code has no parallel in
DegreeWorks. We instead attempt to matching using the name of the degree
type e.g. B.S. or Ph.D. This is currently successful in all cases except
one, where there is disagreement over the correct abbreviation for the
Masters in Management.

Some entries are marked no longer active, meaning there would be no
reason to attempt to find them on DegreeWorks. Since the least recently
deactivated major in DegreeWorks is Applied Ecology at the end of
2006-2007, we can safely assume data from majors deactivated before 2006
won't be found.

## other stuff

The distinction between undergraduate and postgraduate "majors" has been
removed since it does not serve a purpose right now. It can be re-added
if it is determined to be relevant to a different task.

In pursuit of #167, this PR is also a limited pilot of Zod v4, which is
used to validate responses from the aforementioned table. In the future
we might want to validate DegreeWorks responses too.

# Results

Previously, we were brute-forcing valid combinations of the major code
and degree type. This is obviously very inefficient. The number of
requests necessary to obtain all valid undergraduate and postgraduate
programs has been reduced from 7169 to 396 (on the 2024-2025 catalogue),
which strongly correlates to a proportionate time speedup. The
proportion of successful requests is increased from 4.5% to 77%.

It has been verified by external analysis of logs that all (major code,
degree type) tuples successfully fetched by the current scraper but not
the proposed one are misconfigured and do not correspond to meaningful
data.