fix(degreeworks-scraper): infer specialization-major association (#218)
Closes #209.
See the description there for our motivation.

Instead of looking at the list of specializations in a major (which is
no longer consistently provided), we look at the list of overall
specializations and try to match each one to a major.
Some specializations do not match with any major at all. The only way to
resolve this is to attempt every combination of specialization and
major; valid combinations return the specialization block.

Upon analysis, there does seem to be some method to the madness.
The previous soft convention associating specialization code with their
respective major code still appears intact.

This is also progress toward #210 because the list of specializations
previously provided was given as advice when a specialization was not
selected for a major which requires one. In essence, we were previously
totally unaware of optional specializations, but we now scrape and store
them, resulting in the "discovery" of 25 more specializations. We're
still missing the part where we figure out whether specializations are
optional.

In case the aforementioned bruteforcing (over hundreds of majors) is
necessary in the future, and for the benefit of allowing partial
scraping, specialization associations are cached to disk per catalogue
year (it is assumed that these associations won't change per catalogue
year). Any attempt at finding the major associated with a specialization
(including concluding that there is no associated major) is skipped if
the result of an earlier attempt is already in the cache.

The name associated with a specialization in the database has changed to
its name in the specializations mapping, instead of the title of its
block. This is done primarily so that the name given to *e.g.*
PeterPortal agrees with the name seen in the DegreeWorks dropdown menu.

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.