feat(programs): school requirements except humanities (#222)
Closes #131.

Building on previous work on this branch, we now expressly request
requirements from DegreeWorks for its notion of a "college", which
corresponds to schools at UCI.
Similarly to majors and specializations, this *may* lead to an inclusion
of a college block in the response, depending on whether the request was
valid and has applicable college/school requirements.

This block is not identical across all majors; it is allowed in the
DegreeWorks system to change in response to a different choice of major
within the same school.
Therefore a system is also introduced to collect all observed school
blocks in one scrape and test for deep strict (non-coercing) equality
using the `assert` module in Node.
A new table `college_requirement` stores these blocks and a new nullable
column of the same name on `major` may point into a UUID primary key in
that new table.

School requirements are served under the new nullable field on returned
majors, `schoolRequirements`.
OpenAPI and GraphQL types are adjusted accordingly.

Further work is needed in #220.

# testing

REST works:

```
# bs in biological sciences imposes school requirements:
> curl --silent "localhost:8787/v2/rest/programs/major?programId=BS-120" | jq ".data.schoolRequirements" | wc -l
195

# bs in computer science does not (because the school of ICS does not):
> curl --silent "localhost:8787/v2/rest/programs/major?programId=BS-201" | jq ".data.schoolRequirements"
null
```
GraphQL works:


```
> echo 'query Major {
   major(query: { programId: "BS-120" }) {
       id
       name
       schoolRequirements {
           name
       }
   }
}
' \
| tr -d '\n' \
| jq --raw-input --raw-output --compact-output "{query: .}" \
| curl -s -X POST --data @- "http://localhost:8787/v2/graphql" -H "Content-Type: application/json" \
| jq ".data.major"

{
  "id": "BS-120",
  "name": "Major in Biological Sciences",
  "schoolRequirements": {
    "name": "Requirements for School of Biological Sciences"
  }
}

> echo 'query Major {
   major(query: { programId: "BS-201" }) {
       id
       name
       schoolRequirements {
           name
       }
   }
}
' \
| tr -d '\n' \
| jq --raw-input --raw-output --compact-output "{query: .}" \
| curl -s -X POST --data @- "http://localhost:8787/v2/graphql" -H "Content-Type: application/json" \
| jq ".data.major"

{
  "id": "BS-201",
  "name": "Major in Computer Science",
  "schoolRequirements": null
}
```

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [x] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.