Fixed non-breaking misuses of terms division and course level (#250)
## Description

Refactored several files to replace the incorrect usage of the term
division (which should refer to undergraduate and postgraduate) with
course level when it refers to lower division/upper division/graduate
level courses.

## Related Issue

Related to issue #190 

## Motivation and Context

This change was made to increase the consistency and accuracy of term
usage within the code base.

## How Has This Been Tested?

Tested in local environment through WebStorm and Postgres.

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.