feat: /websoc/departments (#252)
## Description

A new endpoint and GraphQL resource are introduced to get a list of all
valid department codes and names.
Per the linked issue, optional parameters also allow for restriction
through a "since" point in time.
Departments which have not been seen in WebSoc since that year or year
and term will not be returned.

## Related Issue

Closes #236.

## Motivation and Context

the years of hardcoding depts are over

## How Has This Been Tested?

curl

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.