feat: Add websoc departments until (#261)
## Description
Add "untilYear" and "untilQuarter" parameters to websoc departments
endpoint.

## Related Issue
Issue #256 

## Motivation and Context
"We'd like to add an "until" pair of parameters as a complement to the
"since" pair, so that collection of department entities for the result
can be restricted in the future direction as well.
This enables other applications like comparison of department offerings
between different terms (if "until" and "since" are set to the same
term) or preventing departments from terms in the future from being
included."

## How Has This Been Tested?
Ran endpoint.

## Screenshots (if appropriate):

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.